refactor(core): start refactoring pbs code
