+++ tfhe/src/core_crypto/commons/math/fft64/x86.rs
@@ -14,9 +14,9 @@ use core::arch::x86::*;
 #[cfg(target_arch = "x86_64")]
 use core::arch::x86_64::*;
 
-use super::super::super::c64;
 use super::TwistiesView;
 use crate::core_crypto::commons::utils::izip;
+use concrete_fft::c64;
 
 use pulp::x86::V3;
 #[cfg(feature = "nightly-avx512")]
@@ -1089,10 +1089,7 @@ pub fn convert_add_backward_torus_u64(
 
 #[cfg(test)]
 mod tests {
-    use crate::core_crypto::fft_impl::fft64::math::fft::{
-        convert_add_backward_torus_scalar, Twisties,
-    };
-
+    use super::super::{convert_add_backward_torus_scalar, Twisties};
     use super::*;
 
     #[test]
