+++ tfhe/src/core_crypto/commons/math/mod.rs
@@ -1,5 +1,6 @@
 //! A module containing general mathematical tools.
 
 pub mod decomposition;
+pub mod fft64;
 pub mod random;
 pub mod torus;
