--- tfhe/src/core_crypto/fft_impl/fft64/crypto/mod.rs
@@ -1,6 +1,0 @@
-pub mod bootstrap;
-pub mod ggsw;
-pub mod wop_pbs;
-
-#[cfg(test)]
-pub mod tests;
