--- tfhe/src/core_crypto/fft_impl/fft64/math/mod.rs
@@ -1,3 +1,0 @@
-pub mod decomposition;
-pub mod fft;
-pub mod polynomial;
