--- tfhe/src/core_crypto/fft_impl/fft64/mod.rs
@@ -1,5 +1,0 @@
-#![doc(hidden)]
-pub use concrete_fft::c64;
-
-pub mod crypto;
-pub mod math;
