+++ tfhe/src/core_crypto/fft_impl/mod.rs
@@ -1,9 +1,4 @@
 pub mod common;
 
-// TODO REFACTOR
-// For now this module is not refactored, it contains high performance code and will be refactored
-// at a later stage. It is self contained, allowing to put it in its own module in the meantime.
-pub mod fft64;
-
 pub mod fft128;
 mod fft128_u128;
