+++ tfhe/src/core_crypto/prelude.rs
@@ -10,10 +10,10 @@ pub use super::commons::computation_buffers::ComputationBuffers;
 pub use super::commons::dispersion::*;
 pub use super::commons::generators::{EncryptionRandomGenerator, SecretRandomGenerator};
 pub use super::commons::math::decomposition::SignedDecomposer;
+pub use super::commons::math::fft64::Fft;
 pub use super::commons::math::random::ActivatedRandomGenerator;
 pub use super::commons::parameters::*;
 pub use super::commons::traits::*;
 pub use super::entities::*;
 pub use super::fft_impl::fft128::math::fft::Fft128;
-pub use super::fft_impl::fft64::math::fft::Fft;
 pub use super::seeders::*;
